+++ app/src/main/java/com/simplepeople/watcha/ui/appscreen/search/SearchScreen.kt
@@ -32,7 +32,7 @@ import kotlinx.coroutines.flow.distinctUntilChanged
 @OptIn(FlowPreview::class)
 @Composable
 fun SearchScreen(
-    navigateToMovieDetails: (Int) -> Unit,
+    navigateToMovieDetails: (Long) -> Unit,
     searchViewModel: SearchViewModel = hiltViewModel()
 ) {
